# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "API documentation for the Laravel application",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Dynamic Server URL from APP_URL"
        }
    ],
    "paths": {
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "List all articles",
                "description": "Retrieve all articles with pagination",
                "operationId": "5f0506d5c66363604ff40841e37502ea",
                "responses": {
                    "200": {
                        "description": "List of articles retrieved successfully"
                    }
                }
            }
        },
        "/articles/search": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Search articles",
                "description": "Search for articles using filters like keyword, date, category, and source",
                "operationId": "ceae3876089eda3586afeafa2bd72d5e",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search keyword",
                        "required": false
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date filter (YYYY-MM-DD)",
                        "required": false
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category filter",
                        "required": false
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Source filter",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filtered articles list retrieved"
                    }
                }
            }
        },
        "/articles/{uuid}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get article by UUID",
                "description": "Retrieve a specific article by its UUID",
                "operationId": "9f6b218e57eb5a65bb709567b6c65e73",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the article",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article retrieved successfully"
                    },
                    "404": {
                        "description": "Article not found"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "description": "Register a user by providing name, email, and password",
                "operationId": "664a650300b0c145840085ccbdb430e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "Password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log in a user",
                "description": "Authenticate user with email and password",
                "operationId": "321117c8cc5ea2a52d22d6607de01080",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log out a user",
                "description": "Logout a user and delete their API token",
                "operationId": "2398725ddafd868354d0df6da04437cf",
                "responses": {
                    "200": {
                        "description": "User logged out successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/password-reset": {
            "post": {
                "tags": [
                    "Password Reset"
                ],
                "summary": "Request password reset link",
                "description": "Send a password reset link to the user's email",
                "operationId": "5eb0c989007d159c1ec70a5dc76e0f43",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset link sent"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/password-reset/confirm": {
            "post": {
                "tags": [
                    "Password Reset"
                ],
                "summary": "Reset user password",
                "description": "Reset password using a valid token",
                "operationId": "80bad5af212b0739359d552eab856439",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "abcdef123456"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "NewPassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "NewPassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/user/preferences": {
            "get": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Get user preferences",
                "description": "Retrieve user preferences for sources, categories, and authors",
                "operationId": "11e44beda1072854089eb93973a06902",
                "responses": {
                    "200": {
                        "description": "User preferences retrieved successfully"
                    },
                    "404": {
                        "description": "Preferences not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Set user preferences",
                "description": "Save user preferences for sources, categories, and authors",
                "operationId": "60b5a61e7efc095204ffe08be5838fa0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sources": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "TechCrunch"
                                        }
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "Technology"
                                        }
                                    },
                                    "authors": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "John Doe"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preferences saved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/user/personalized-feed": {
            "get": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Get personalized feed",
                "description": "Retrieve personalized feed based on user preferences",
                "operationId": "20efae9905e89338e392b6d2e28910a5",
                "responses": {
                    "200": {
                        "description": "Personalized feed retrieved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Articles",
            "description": "Articles"
        },
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Password Reset",
            "description": "Password Reset"
        },
        {
            "name": "User Preferences",
            "description": "User Preferences"
        }
    ]
}